@@ -4,12 +4,8 @@ import { Router } from "@angular/router";
 import { Action, Store } from "@ngrx/store";
 import { Effect, Actions } from "@ngrx/effects";
 import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/withLatestFrom';
-import 'rxjs/add/operator/do';
+import { of } from "rxjs/observable/of";
+import { map, switchMap, catchError, withLatestFrom, tap } from "rxjs/operators";
 
 import * as contact from './../actions/employee-contact.actions';
 import * as fromRoot from './../../reducers';
@@ -33,41 +29,54 @@ export class EmployeeContactInformationEffects {
     @Effect()
         getContact$ : Observable<Action> = this._actions$
         .ofType(contact.GET_CONTACT)
-        .withLatestFrom(
-            this._store$.select(fromRoot.getEmployeeId),
-        )
-        .switchMap( ([action, employeeId]) => {
-            return this._service.getContact(employeeId)
-            .map((response) => new contact.GetContactSuccess(response.data) )
-            .catch(err => Observable.of(new contact.GetContactError(err) ))
-
-        });
+        .pipe(
+            withLatestFrom(
+                this._store$.select(fromRoot.getEmployeeId)
+            ),
+            switchMap( ([action, employeeId]) => {
+                return this._service.getContact(employeeId)
+                .pipe(
+                    map((response) => new contact.GetContactSuccess(response.data) ),
+                    catchError(err => of(new contact.GetContactError(err) ))
+                )
+            })
+        );
+       
     @Effect({dispatch : false})
         error$ = this._actions$
                  .ofType<contact.GetContactError>(contact.GET_CONTACT_ERROR)
-                 .map((action) => action.payload)
-                 .do((payload) => { this._toastr.errorHandler(payload)})
+                 .pipe(
+                    map((action) => action.payload),
+                    tap((payload) => { this._toastr.errorHandler(payload)})
+                 );
+                    
 
 
     @Effect()
         save$ : Observable<Action> = this._actions$
         .ofType<contact.SaveContact>(contact.SAVE_CONTACT)
-        .map( (action) => action.payload)
-        .switchMap((payload) => {
-
-            this._loader.openDialog();
-            
-            return this._service.updateContact(payload)
-                .map((response) =>  new contact.SaveContactSuccess(payload) )
-                .catch((err) => Observable.of( new contact.GetContactError(err) ))
-                .do(() => this._loader.closeDialog())
-
-        })
+        .pipe(
+            map( (action) => action.payload),
+            switchMap((payload) => {
+    
+                this._loader.openDialog();
+                
+                return this._service.updateContact(payload)
+                .pipe(
+                    map((response) =>  new contact.SaveContactSuccess(payload) ),
+                    catchError((err) => of( new contact.GetContactError(err) )),
+                    tap(() => this._loader.closeDialog())
+                )
+            })
+        );
+        
 
     @Effect({dispatch : false})
         saveSuccess$  = this._actions$
                        .ofType(contact.SAVE_CONTACT_SUCCESS)
-                       .do(() => { this._toastr.saveSuccess(); })
-                       
+                       .pipe(
+                            tap(() => { this._toastr.saveSuccess(); })
+                       );
+               
                                  
 }
