@@ -4,12 +4,8 @@ import { Router } from "@angular/router";
 import { Action, Store } from "@ngrx/store";
 import { Effect, Actions } from "@ngrx/effects";
 import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/withLatestFrom';
-import 'rxjs/add/operator/do';
+import { of } from "rxjs/observable/of";
+import { map, switchMap, catchError, withLatestFrom, tap } from "rxjs/operators";
 
 import * as empEmployment from './../actions/employee-employment.actions';
 import * as fromRoot from './../../reducers';
@@ -33,41 +29,56 @@ export class EmployeeEmploymentInformationEffects {
     @Effect()
         getEmployment$ : Observable<Action> = this._actions$
         .ofType(empEmployment.GET_EMPLOYMENT)
-        .withLatestFrom(
-            this._store$.select(fromRoot.getEmployeeId),
-        )
-        .switchMap( ([action, employeeId]) => {
-            return this._service.getEmployment(employeeId)
-            .map((response) => new empEmployment.GetEmploymentSuccess(response.data) )
-            .catch(err => Observable.of(new empEmployment.GetEmploymentError(err) ))
-            
-        });
+        .pipe(
+            withLatestFrom(
+                this._store$.select(fromRoot.getEmployeeId)
+            ),
+            switchMap( ([action, employeeId]) => {
+                return this._service.getEmployment(employeeId)
+                .pipe(
+                    map((response) => new empEmployment.GetEmploymentSuccess(response.data) ),
+                    catchError(err => of(new empEmployment.GetEmploymentError(err) ))
+                )
+                
+            })
+        );
+        
     @Effect({dispatch : false})
         error$ = this._actions$
                  .ofType<empEmployment.GetEmploymentError>(empEmployment.GET_EMPLOYMENT_ERROR)
-                 .map((action) => action.payload)
-                 .do((payload) => { this._toastr.errorHandler(payload)})
+                 .pipe(
+                    map((action) => action.payload),
+                    tap((payload) => { this._toastr.errorHandler(payload)})
+                 )
 
 
     @Effect()
         save$ : Observable<Action> = this._actions$
         .ofType<empEmployment.SaveEmployment>(empEmployment.SAVE_EMPLOYMENT)
-        .map( (action) => action.payload)
-        .switchMap((payload) => {
-
-            this._loader.openDialog();
-            
-            return this._service.updateEmployment(payload)
-                .map((response) =>  new empEmployment.SaveEmploymentSuccess(payload) )
-                .catch((err) => Observable.of( new empEmployment.GetEmploymentError(err) ))
-                .do(() => this._loader.closeDialog())
-
-        })
+        .pipe(
+            map( (action) => action.payload),
+            switchMap((payload) => {
+    
+                this._loader.openDialog();
+                
+                return this._service.updateEmployment(payload)
+                .pipe(
+                    map((response) =>  new empEmployment.SaveEmploymentSuccess(payload) ),
+                    catchError((err) => of( new empEmployment.GetEmploymentError(err) )),
+                    tap(() => this._loader.closeDialog())
+                )
+    
+            })
+        );
+        
 
     @Effect({dispatch : false})
         saveSuccess$  = this._actions$
                        .ofType(empEmployment.SAVE_EMPLOYMENT_SUCCESS)
-                       .do(() => { this._toastr.saveSuccess(); })
+                       .pipe(
+                            tap(() => { this._toastr.saveSuccess(); })
+                       );
+                       
                        
                                  
 }
