@@ -4,13 +4,8 @@ import { Router } from "@angular/router";
 import { Action, Store } from "@ngrx/store";
 import { Effect, Actions } from "@ngrx/effects";
 import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/withLatestFrom';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/filter';
+import { of } from "rxjs/observable/of";
+import { map, switchMap, catchError, withLatestFrom, tap, filter } from "rxjs/operators";
 
 import * as empPersonal from './../actions/employee-personal.actions';
 import * as fromRoot from './../../reducers';
@@ -34,43 +29,56 @@ export class EmployeePersonalInformationEffects {
     @Effect()
         getPersonal$ : Observable<Action> = this._actions$
         .ofType(empPersonal.GET_PERSONAL)
-        .withLatestFrom(
-            this._store$.select(fromRoot.getEmployeeId),
-        )
-        .switchMap( ([action, employeeId]) => {
-            return this._service.getPersonal(employeeId)
-            .map((response) => new empPersonal.GetPersonalSuccess(response.data) )
-            .catch(err => Observable.of(new empPersonal.GetPersonalError(err) ))
-
-        });
+        .pipe(
+            withLatestFrom(
+                this._store$.select(fromRoot.getEmployeeId)
+            ),
+            switchMap( ([action, employeeId]) => {
+                return this._service.getPersonal(employeeId)
+                .pipe(
+                    map((response) => new empPersonal.GetPersonalSuccess(response.data) ),
+                    catchError(err => of(new empPersonal.GetPersonalError(err) ))
+                )
+    
+            })
+        );
+        
     @Effect({dispatch : false})
         error$ = this._actions$
                  .ofType<empPersonal.GetPersonalError>(empPersonal.GET_PERSONAL_ERROR)
-                 .map((action) => action.payload)
-                 .do((payload) => { this._toastr.errorHandler(payload)})
-                 .filter(payload => payload.status == 404)
-                 .do(() => this._router.navigateByUrl('404/page-not-found'))
+                 .pipe(
+                    map((action) => action.payload),
+                    tap((payload) => { this._toastr.errorHandler(payload)}),
+                    filter(payload => payload.status == 404),
+                    tap(() => this._router.navigateByUrl('404/page-not-found'))
+                 );
 
 
     @Effect()
         save$ : Observable<Action> = this._actions$
         .ofType<empPersonal.SavePersonal>(empPersonal.SAVE_PERSONAL)
-        .map( (action) => action.payload)
-        .switchMap((payload) => {
+        .pipe(
+            map( (action) => action.payload),
+            switchMap((payload) => {
 
-            this._loader.openDialog();
-            
-            return this._service.updatePersonal(payload)
-                .map((response) =>  new empPersonal.SavePersonalSuccess(payload) )
-                .catch((err) => Observable.of( new empPersonal.GetPersonalError(err) ))
-                .do(() => this._loader.closeDialog())
+                this._loader.openDialog();
+                
+                return this._service.updatePersonal(payload)
+                .pipe(
+                    map((response) =>  new empPersonal.SavePersonalSuccess(payload) ),
+                    catchError((err) => of( new empPersonal.GetPersonalError(err) )),
+                    tap(() => this._loader.closeDialog())
+                )
 
-        })
+            })
+        );
 
     @Effect({dispatch : false})
         saveSuccess$  = this._actions$
                        .ofType(empPersonal.SAVE_PERSONAL_SUCCESS)
-                       .do(() => { this._toastr.saveSuccess(); })
+                       .pipe(
+                            tap(() => { this._toastr.saveSuccess(); })
+                       );
                        
                                  
 }
