@@ -5,6 +5,6 @@
 
 export const environment = {
   production: false,
-  endPoint : 'http://127.0.0.1:8000'
-  //endPoint : 'http://staticscript.com/ipm3-laravel-api/public'
+  //endPoint : 'http://127.0.0.1:8000'
+  endPoint : 'http://staticscript.com/ipm3-laravel-api/public'
 };
